requirements.txt
asgiref==3.3.1
Django==3.1.7
django-ninja==0.11.0
pydantic==1.7.3
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
